refactor: reorganize redux
