+++ src/store/reducers/index.ts
@@ -1,0 +1,8 @@
+import { combineReducers } from "redux";
+import { websiteBuilderReducer } from "./websiteBuilderReducer";
+
+const rootReducer = combineReducers({
+  websiteBuilder: websiteBuilderReducer,
+});
+
+export default rootReducer;
